Changed Functors/Monads/Either folder topology
