+++ src/Functors/Monads/Either/index.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * either monad functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/Either.php';
+require_once __DIR__ . '/FromLeft.php';
+require_once __DIR__ . '/FromRight.php';
+require_once __DIR__ . '/IsLeft.php';
+require_once __DIR__ . '/IsRight.php';
+require_once __DIR__ . '/Lefts.php';
+require_once __DIR__ . '/PartitionEithers.php';
+require_once __DIR__ . '/Rights.php';
